// solauto-sdk/src/generated/types/priceType.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { Serializer, scalarEnum } from '@metaplex-foundation/umi/serializers';

export enum PriceType {
  Realtime,
  Ema,
}

export type PriceTypeArgs = PriceType;

export function getPriceTypeSerializer(): Serializer<PriceTypeArgs, PriceType> {
  return scalarEnum<PriceType>(PriceType, {
    description: 'PriceType',
  }) as Serializer<PriceTypeArgs, PriceType>;
}