// solauto-sdk/src/generated/types/tokenBalanceChange.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import {
  Serializer,
  array,
  struct,
  u64,
  u8,
} from '@metaplex-foundation/umi/serializers';
import {
  TokenBalanceChangeType,
  TokenBalanceChangeTypeArgs,
  getTokenBalanceChangeTypeSerializer,
} from '.';

export type TokenBalanceChange = {
  changeType: TokenBalanceChangeType;
  padding1: Array<number>;
  amountUsd: bigint;
};

export type TokenBalanceChangeArgs = {
  changeType: TokenBalanceChangeTypeArgs;
  padding1: Array<number>;
  amountUsd: number | bigint;
};

export function getTokenBalanceChangeSerializer(): Serializer<
  TokenBalanceChangeArgs,
  TokenBalanceChange
> {
  return struct<TokenBalanceChange>(
    [
      ['changeType', getTokenBalanceChangeTypeSerializer()],
      ['padding1', array(u8(), { size: 7 })],
      ['amountUsd', u64()],
    ],
    { description: 'TokenBalanceChange' }
  ) as Serializer<TokenBalanceChangeArgs, TokenBalanceChange>;
}